refactor: included PlaceViewModel at main activity to use list of Place across fragments. moved PlaceRepository api fetch to PlaceViewModel to fetch places.
