+++ app/src/main/java/team2/kakigowhere/MainActivity.kt
@@ -2,14 +2,18 @@ package team2.kakigowhere
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
+import team2.kakigowhere.data.model.PlaceViewModel
 
 class MainActivity : AppCompatActivity() {
+
+    private val placeViewModel: PlaceViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -21,6 +25,9 @@ class MainActivity : AppCompatActivity() {
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, 0) // no bottom padding to account for bottom nav bar
             insets
         }
+
+        // make api call to get all places
+        placeViewModel.loadPlaces()
 
         // sets up navigation host and find navigation controller
         val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
