+++ app/src/main/java/team2/kakigowhere/data/api/ApiService.kt
@@ -7,25 +7,31 @@ import retrofit2.http.Header
 import retrofit2.http.POST
 import retrofit2.http.Path
 import retrofit2.http.Query
-import team2.kakigowhere.data.model.Place
+import team2.kakigowhere.data.model.CreateItineraryDTO
+import team2.kakigowhere.data.model.ItineraryDTO
+import team2.kakigowhere.data.model.ItineraryDetailDTO
 import team2.kakigowhere.data.model.PlaceDTO
 import team2.kakigowhere.data.model.RatingItem
 import team2.kakigowhere.data.model.RatingRequest
 import team2.kakigowhere.data.model.RatingSummary
-import team2.kakigowhere.data.model.CreateItineraryDTO
-import team2.kakigowhere.data.model.ItineraryDTO
-import team2.kakigowhere.data.model.ItineraryDetailDTO
 
 interface ApiService {
+
+    // places api
     @GET("places")
     suspend fun getPlaces(): Response<List<PlaceDTO>>
 
     @GET("places/id/{placeId}")
-    suspend fun getPlaceDetail(@Path("placeId") placeId: Long): Response<PlaceDTO>
+    suspend fun getPlaceDetail(
+        @Path("placeId") placeId: Long
+    ): Response<PlaceDTO>
 
     @GET("ratings/{placeId}/summary")
-    suspend fun getRatingSummary(@Path("placeId") placeId: Long): Response<RatingSummary>
+    suspend fun getRatingSummary(
+        @Path("placeId") placeId: Long
+    ): Response<RatingSummary>
 
+    // tourist api
     // may be empty as Tourist may not have given rating
     @GET("ratings/{placeId}/me")
     suspend fun getMyRating(
@@ -46,6 +52,7 @@ interface ApiService {
         @Body request: RatingRequest
     ): Response<RatingItem>
 
+    // itinerary api
     @GET("itinerary/{email}")
     suspend fun getItineraries(
         @Path("email") email: String
